requirements.txt
jinja2
fastapi
pillow
torch
uvicorn
pydantic
torchvision
torch
matplotlib